requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.0.4
django-filter==21.1
django-pandas==0.6.6
djangorestframework==3.13.1
executing==0.8.3
greenlet==1.1.2
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.29
mysqlclient==2.1.0
numpy==1.22.4
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
protobuf==4.21.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
traitlets==5.2.2.post1
wcwidth==0.2.5
